refactor: 불필요한 폴더 제거
