+++ src/components/common/icons/WriteIcon.tsx
